+++ src/Arborescence.Models/Adjacency/ListAdjacencyGraph_2.cs
@@ -29,9 +29,9 @@ namespace Arborescence.Models
         internal ListAdjacencyGraph(TVertexMultimap neighborsByVertex) => _neighborsByVertex = neighborsByVertex;
 
         /// <summary>
-        /// Gets the number of vertices.
+        /// Gets the number of vertices that can have out-neighbors.
         /// </summary>
-        public int VertexCount
+        public int TailCount
         {
             get
             {