+++ src/Arborescence.Models/Incidence/ListIncidenceGraph_4.cs
@@ -40,9 +40,9 @@ namespace Arborescence.Models
         }
 
         /// <summary>
-        /// Gets the number of vertices.
+        /// Gets the number of vertices that can have out-edges.
         /// </summary>
-        public int VertexCount
+        public int TailCount
         {
             get
             {