+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.16.5] - 2024-01-07
+
+### Added
+
+- Primitives: `EqualityComparerEquatable<T, TComparer>`, `DefaultEqualityComparerEquatable<T>`
+- Primitives.Specialized: Another overload of `Int32ReadOnlyDictionary<TValue>.CreateWithAbsence<TValueList, TComparer>()`
+
+### Changed
+
+- Models: Renamed `VertexCount` to `TailCount` in some graph types.
+- Primitives.Specialized: Changed signature of `Int32ReadOnlyDictionary<TValue>.CreateWithAbsence<TValueList>()` to use `DefaultComparerEquatable<TValue?>` instead of `DefaultAbsence<T>`.
+- Primitives.Specialized: Changed signature of `Int32IndirectReadOnlyDictionary<TKey, TValue>.CreateFromListWithAbsence<TKeyToIndexMap, TValueList>()` to use `DefaultComparerEquatable<TValue?>` instead of `DefaultAbsence<T>`.
+
+### Removed
+
+- Primitives: `DefaultAbsence<T>`
+- Primitives.Specialized: `ArrayPrefix<T>`
+
 ## [0.16.4] - 2024-01-02
 
 ### Changed
@@ -16,7 +34,7 @@
 ### Added
 
 - Models: `net8.0` target framework moniker to support `FrozenSet<T>`
-- Models:  `FrozenSetEnumeratorProvider<T>`, `HashSetEnumeratorProvider<T>`
+- Models: `FrozenSetEnumeratorProvider<T>`, `HashSetEnumeratorProvider<T>`
 - Models: `ArrayEnumeratorProvider<T>`, `ArraySegmentEnumeratorProvider<T>`
 - Models: `ReadOnlyAdjacencyGraphFactory<TVertex>`, `ReadOnlyIncidenceGraphFactory<TVertex, TEdge>`
 
@@ -455,7 +473,9 @@
 
 - Abstractions: The interface for graphs to be examined in a data-structure agnostic fashion.
 
-[Unreleased]: https://github.com/qbit86/arborescence/compare/arborescence-0.16.4...HEAD
+[Unreleased]: https://github.com/qbit86/arborescence/compare/arborescence-0.16.5...HEAD
+
+[0.16.5]: https://github.com/qbit86/arborescence/compare/arborescence-0.16.4...arborescence-0.16.5
 
 [0.16.4]: https://github.com/qbit86/arborescence/compare/arborescence-0.16.3...arborescence-0.16.4